Show keyboard shortcut for blank button - Move keys.tsx to components/keys.tsx - Add className and size props to all keys
