+++ packages/scrabble-solver/src/components/Board/components/ToggleDirectionButton/ToggleDirectionButton.module.scss
@@ -22,5 +22,5 @@
 
 .arrows {
   display: inline-block;
-  margin-left: var(--spacing--s);
+  margin-left: var(--spacing--m);
 }
